Moving trace id generation to the config class
